Add files via upload
This is an Arduino example sketch.  If you run it on an Arduino Micro and connect a potentiometer to pins A0 and A1, they will show up as joystick axis in the drive station.  

These will be our trim knobs that the operator can use to manually offset the elevator height and swingarm angle if it is needed mid match.

Note:  The elevator trim should NEVER allow the elevator setpoint to go past the maximum elevator height. We will need to add error checking to the elevator setpoint to make sure thst doesn't happen.